requirements.txt
pytest==7.2.2
selenium
pytest-html
pytest-xdist
webdriver-manager
openpyxl
allure-pytest
mysql-connector-python
